12.Refactor Special Numbers.java
You are given a working code that is a solution to Problem 9. Special Numbers. However, the variables are improperly named, declared before they are needed and some of them are used for multiple things. Without using your previous solution, modify the code so that it is easy to read and understand.
Hints
•	Reduce the span of the variables by declaring them in the moment they receive a value, not before
•	Rename your variables to represent their real purpose (example: "toe" should become isSpecialNum, etc.)
•	Search for variables that have multiple purpose. If you find any, introduce a new variable